관리자 목록 진행 중

+++ src/main/java/com/bugjeogbugjeog/app/bugjeogbugjeog/domain/dao/InquiryBoardDAO.java
@@ -105,7 +105,7 @@ public class InquiryBoardDAO {
     }
 
     // 문의 카운트
-    public int count() {
+    public Long count() {
         return inquiryBoardMapper.count();
     }
